# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: Mental Social API
  version: 1.0.0
  description: |
    Express + SQLite + Sequelize backend for mini social network + mood tracker.
servers:
  - url: http://localhost:4000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
              example: Validation error
            details:
              oneOf:
                - type: array
                  items:
                    type: object
                    properties:
                      type: { type: string, example: field }
                      msg: { type: string, example: Invalid value }
                      path: { type: string, example: email }
                      location: { type: string, example: body }
                - type: object
                  additionalProperties: true
          required: [message]
      required: [error]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
      required: [ok]
...
    UserPublic:
      type: object
      properties:
        id: { type: integer, example: 1 }
        username: { type: string, example: ivan }
        displayName: { type: string, nullable: true, example: ivan }
        avatarUrl: { type: string, nullable: true, example: /uploads/abc.png }
      required: [id, username]
    UserPrivate:
      allOf:
        - $ref: '#/components/schemas/UserPublic'
        - type: object
          properties:
            email: { type: string, example: ivan@test.com }
            bio: { type: string, nullable: true, example: "hi" }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
          required: [email, createdAt, updatedAt]
    AuthResponse:
      type: object
      properties:
        token: { type: string, example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." }
        user:
          $ref: '#/components/schemas/UserPrivate'
      required: [token, user]

    # ===== Mood Entries =====
...
    Comment:
      type: object
      properties:
        id: { type: integer, example: 100 }
        postId: { type: integer, example: 55 }
        authorId: { type: integer, example: 2 }
        text: { type: string, example: "Nice!" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        author:
          allOf:
            - $ref: '#/components/schemas/UserPublic'
          nullable: true
      required: [id, postId, authorId, text, createdAt, updatedAt]
    CommentCreate:
      type: object
      properties:
        text: { type: string, example: "держись" }
      required: [text]
    CommentUpdate:
      type: object
      properties:
        text: { type: string, example: "updated text" }
...
    Friendship:
      type: object
      properties:
        id: { type: integer, example: 77 }
        requesterId: { type: integer, example: 1 }
        addresseeId: { type: integer, example: 2 }
        status:
          type: string
          enum: [pending, accepted, blocked]
          example: pending
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        requester:
          allOf:
            - $ref: '#/components/schemas/UserPublic'
          nullable: true
        addressee:
          allOf:
            - $ref: '#/components/schemas/UserPublic'
          nullable: true
      required: [id, requesterId, addresseeId, status, createdAt, updatedAt]
    FriendshipCreate:
      type: object
      properties:
        addresseeId: { type: integer, example: 2 }
      required: [addresseeId]
    FriendshipUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, accepted, blocked]
          example: accepted
      required: [status]
    FriendshipListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Friendship' }
      required: [data]
    FriendshipSingleResponse:
      type: object
      properties:
        data: { $ref: '#/components/schemas/Friendship' }
      required: [data]

    # ===== Uploads =====
...
  responses:
    Unauthorized:
      description: Unauthorized (missing/invalid Bearer token)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            missingToken:
              value:
                error: { message: "Missing Bearer token" }
            invalidToken:
              value:
                error: { message: "Invalid token" }
    Forbidden:
      description: Forbidden (ownership / permissions)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            forbidden:
              value:
                error: { message: "Forbidden" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            nf:
              value:
                error: { message: "Not found" }
    ValidationError:
      description: Validation error (express-validator)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user (returns JWT)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username, password]
              properties:
                email: { type: string, format: email, example: ivan@test.com }
                username: { type: string, example: ivan }
                password: { type: string, example: Password123! }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '409':
          description: User already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                exists:
                  value:
                    error: { message: "User already exists" }
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email or username (identifier) (returns JWT)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string, example: ivan@test.com }
                password: { type: string, example: Password123! }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                invalid:
                  value:
                    error: { message: "Invalid credentials" }
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user by token
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/UserPrivate' }
                required: [user]
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===== Users =====
...
  /api/posts/{postId}/comments:
    post:
      tags: [Comments]
      summary: Create comment for a post (author is current user)
      parameters:
        - in: path
          name: postId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommentCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Comment' }
                required: [data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Post not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                nf:
                  value:
                    error: { message: "Post not found" }
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/comments/{id}:
    put:
      tags: [Comments]
      summary: Update my comment (ownership enforced)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommentUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Comment' }
                required: [data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Comments]
      summary: Delete my comment (ownership enforced)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===== Friendships =====
  /api/friendships:
    get:
      tags: [Friendships]
      summary: List my friendships (sent + received), optionally filter by status
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, accepted, blocked]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FriendshipListResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/ValidationError'

    post:
      tags: [Friendships]
      summary: Send friend request to addresseeId (creates pending)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FriendshipCreate' }
      responses:
        '201':
          description: Created (or 200 if already exists)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FriendshipSingleResponse' }
        '200':
          description: Already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FriendshipSingleResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                nf:
                  value:
                    error: { message: "User not found" }
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/friendships/{id}:
    put:
      tags: [Friendships]
      summary: Update friendship status (requester OR addressee only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FriendshipUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FriendshipSingleResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Friendships]
      summary: Delete friendship (requester OR addressee only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===== Uploads =====
...